+++ docs/court_wizard.d.ts
@@ -6,14 +6,14 @@ export type InitInput = RequestInfo | URL | Response | BufferSource | WebAssembl
 export interface InitOutput {
   readonly memory: WebAssembly.Memory;
   readonly main: (a: number, b: number) => number;
-  readonly __wasm_bindgen_func_elem_130837: (a: number, b: number, c: number, d: number) => void;
-  readonly __wasm_bindgen_func_elem_130714: (a: number, b: number) => void;
-  readonly __wasm_bindgen_func_elem_126149: (a: number, b: number) => void;
-  readonly __wasm_bindgen_func_elem_126128: (a: number, b: number) => void;
-  readonly __wasm_bindgen_func_elem_130835: (a: number, b: number) => void;
-  readonly __wasm_bindgen_func_elem_130836: (a: number, b: number, c: number) => void;
-  readonly __wasm_bindgen_func_elem_129177: (a: number, b: number, c: number) => void;
-  readonly __wasm_bindgen_func_elem_129167: (a: number, b: number) => void;
+  readonly __wasm_bindgen_func_elem_130854: (a: number, b: number, c: number, d: number) => void;
+  readonly __wasm_bindgen_func_elem_130731: (a: number, b: number) => void;
+  readonly __wasm_bindgen_func_elem_130852: (a: number, b: number) => void;
+  readonly __wasm_bindgen_func_elem_130853: (a: number, b: number, c: number) => void;
+  readonly __wasm_bindgen_func_elem_126166: (a: number, b: number) => void;
+  readonly __wasm_bindgen_func_elem_126145: (a: number, b: number) => void;
+  readonly __wasm_bindgen_func_elem_129194: (a: number, b: number, c: number) => void;
+  readonly __wasm_bindgen_func_elem_129184: (a: number, b: number) => void;
   readonly __wbindgen_export: (a: number, b: number) => number;
   readonly __wbindgen_export2: (a: number, b: number, c: number, d: number) => number;
   readonly __wbindgen_export3: (a: number) => void;