+++ docs/court_wizard_bg.wasm.d.ts
@@ -2,14 +2,14 @@
 /* eslint-disable */
 export const memory: WebAssembly.Memory;
 export const main: (a: number, b: number) => number;
-export const __wasm_bindgen_func_elem_130837: (a: number, b: number, c: number, d: number) => void;
-export const __wasm_bindgen_func_elem_130714: (a: number, b: number) => void;
-export const __wasm_bindgen_func_elem_126149: (a: number, b: number) => void;
-export const __wasm_bindgen_func_elem_126128: (a: number, b: number) => void;
-export const __wasm_bindgen_func_elem_130835: (a: number, b: number) => void;
-export const __wasm_bindgen_func_elem_130836: (a: number, b: number, c: number) => void;
-export const __wasm_bindgen_func_elem_129177: (a: number, b: number, c: number) => void;
-export const __wasm_bindgen_func_elem_129167: (a: number, b: number) => void;
+export const __wasm_bindgen_func_elem_130854: (a: number, b: number, c: number, d: number) => void;
+export const __wasm_bindgen_func_elem_130731: (a: number, b: number) => void;
+export const __wasm_bindgen_func_elem_130852: (a: number, b: number) => void;
+export const __wasm_bindgen_func_elem_130853: (a: number, b: number, c: number) => void;
+export const __wasm_bindgen_func_elem_126166: (a: number, b: number) => void;
+export const __wasm_bindgen_func_elem_126145: (a: number, b: number) => void;
+export const __wasm_bindgen_func_elem_129194: (a: number, b: number, c: number) => void;
+export const __wasm_bindgen_func_elem_129184: (a: number, b: number) => void;
 export const __wbindgen_export: (a: number, b: number) => number;
 export const __wbindgen_export2: (a: number, b: number, c: number, d: number) => number;
 export const __wbindgen_export3: (a: number) => void;